+++ app/build.gradle
@@ -6,7 +6,7 @@ android {
     buildToolsVersion "29.0.3"
 
     defaultConfig {
-        applicationId "com.shukhrat.wbpvp"
+        applicationId "com.piu.ridp"
         minSdkVersion 19
         targetSdkVersion 29
         versionCode 102
