+++ app/src/main/java/com/piu/ridp/ui/news/NewsFragment.java
@@ -1,0 +1,84 @@
+package com.piu.ridp.ui.news;
+
+import android.os.Bundle;
+import android.view.KeyEvent;
+import android.view.LayoutInflater;
+import android.view.MotionEvent;
+import android.view.View;
+import android.view.ViewGroup;
+import android.webkit.WebView;
+import android.webkit.WebViewClient;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
+
+import com.piu.ridp.R;
+
+import org.jetbrains.annotations.NotNull;
+
+public class NewsFragment extends Fragment {
+
+    public NewsFragment() {
+        // Required empty public constructor
+    }
+
+    public View onCreateView(@NonNull LayoutInflater inflater,
+                             ViewGroup container, Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.fragment_news, container, false);
+    }
+
+    @Override
+    public void onViewCreated(@NonNull @NotNull View view, @Nullable @org.jetbrains.annotations.Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        final SwipeRefreshLayout swipeRefreshLayout = view.findViewById(R.id.swipe_refresh);
+        final WebView webView = view.findViewById(R.id.web_view);
+
+        webView.getSettings().setJavaScriptEnabled(true);
+        webView.setWebViewClient(new WebViewClient() {
+            @Override
+            public void onLoadResource(WebView view, String url) {
+                super.onLoadResource(view, url);
+                view.loadUrl("javascript: (function() { " +
+                        "document.getElementsByClassName('header-area')[0].style.display='none';" +
+                        "document.getElementsByClassName('breadcumb-area')[0].style.display='none';" +
+                        "document.getElementsByClassName('footer-area')[0].style.display='none';" +
+                        "document.getElementsByClassName('globalClass_191b')[0].style.display='none';" +
+                        "})()");
+            }
+
+            @Override
+            public boolean shouldOverrideUrlLoading(WebView view, String url) {
+                if (url.contains("jivosite.com")) return false;
+                return super.shouldOverrideUrlLoading(view, url);
+            }
+
+            @Override
+            public void onPageFinished(WebView view, String url) {
+                swipeRefreshLayout.setRefreshing(false);
+                super.onPageFinished(view, url);
+            }
+        });
+        webView.loadUrl("http://ridp.uz/" + getString(R.string.app_lang) + "/news/");
+        webView.setOnKeyListener(new View.OnKeyListener() {
+            @Override
+            public boolean onKey(View v, int keyCode, KeyEvent event) {
+                if (keyCode == KeyEvent.KEYCODE_BACK && event.getAction() == MotionEvent.ACTION_UP && webView.canGoBack()) {
+                    webView.goBack();
+                    return true;
+                }
+                return false;
+            }
+        });
+
+        swipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
+            @Override
+            public void onRefresh() {
+                webView.loadUrl("http://ridp.uz/" + getString(R.string.app_lang) + "/news/");
+            }
+        });
+
+    }
+}
