+++ app/src/main/java/com/piu/ridp/ui/slideshow/SlideshowViewModel.java
@@ -1,0 +1,19 @@
+package com.piu.ridp.ui.slideshow;
+
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
+
+public class SlideshowViewModel extends ViewModel {
+
+    private MutableLiveData<String> mText;
+
+    public SlideshowViewModel() {
+        mText = new MutableLiveData<>();
+        mText.setValue("This is slideshow fragment");
+    }
+
+    public LiveData<String> getText() {
+        return mText;
+    }
+}
