+++ app/src/main/java/com/piu/ridp/ui/sliding_panel/BottomSheetFragment.java
@@ -1,0 +1,34 @@
+package com.piu.ridp.ui.sliding_panel;
+
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment;
+import com.piu.ridp.R;
+
+
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class BottomSheetFragment extends BottomSheetDialogFragment {
+    public BottomSheetFragment() {
+        // Required empty public constructor
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_bottom_sheet_dialog, container, false);
+    }
+}
