--- app/src/main/java/com/shukhrat/wbpvp/authentification/EnterPhoneNumber.java
@@ -1,48 +1,0 @@
-package com.shukhrat.wbpvp.authentification;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.EditText;
-import android.widget.Spinner;
-
-import com.shukhrat.wbpvp.R;
-import com.shukhrat.wbpvp.language.BaseActivity;
-
-public class EnterPhoneNumber extends BaseActivity {
-
-    private Spinner spinner;
-    private EditText editText;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_phone_number);
-
-        editText = findViewById(R.id.editTextMobile);
-
-        findViewById(R.id.buttonContinue).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-
-                String number = editText.getText().toString().trim();
-
-                if (number.isEmpty()) {
-                    editText.setError("Valid number is required");
-                    editText.requestFocus();
-                    return;
-                }
-
-                String phoneNumber = "+998" + number;
-
-                Intent intent = new Intent(EnterPhoneNumber.this, VerifyPhoneActivity.class);
-                intent.putExtra("phonenumber", phoneNumber);
-                startActivity(intent);
-
-            }
-        });
-    }
-}
