--- app/src/main/java/com/shukhrat/wbpvp/language/BaseActivity.java
@@ -1,31 +1,0 @@
-package com.shukhrat.wbpvp.language;
-
-import android.content.Context;
-import android.content.pm.ActivityInfo;
-import android.content.pm.PackageManager;
-import android.os.Bundle;
-import androidx.annotation.Nullable;
-import androidx.appcompat.app.AppCompatActivity;
-
-import static android.content.pm.PackageManager.GET_META_DATA;
-public abstract class BaseActivity extends AppCompatActivity {
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        resetTitles();
-    }
-    @Override
-    protected void attachBaseContext(Context base) {
-        super.attachBaseContext(LocaleManager.setLocale(base));
-    }
-    protected void resetTitles() {
-        try {
-            ActivityInfo info = getPackageManager().getActivityInfo(getComponentName(), GET_META_DATA);
-            if (info.labelRes != 0) {
-                setTitle(info.labelRes);
-            }
-        } catch (PackageManager.NameNotFoundException e) {
-            e.printStackTrace();
-        }
-    }
-}
