--- app/src/main/java/com/shukhrat/wbpvp/pvp.java
@@ -1,38 +1,0 @@
-package com.shukhrat.wbpvp;
-
-import android.app.Application;
-import android.content.Context;
-import android.content.res.Configuration;
-
-import com.google.firebase.database.FirebaseDatabase;
-import com.shukhrat.wbpvp.language.LocaleManager;
-import com.squareup.picasso.OkHttp3Downloader;
-import com.squareup.picasso.Picasso;
-
-public class pvp extends Application {
-
-    @Override
-    public void onCreate() {
-        super.onCreate();
-
-        FirebaseDatabase.getInstance().setPersistenceEnabled(true);
-
-        Picasso.Builder builder = new Picasso.Builder(this);
-        builder.downloader(new OkHttp3Downloader(this, Integer.MAX_VALUE));
-        Picasso build = builder.build();
-        build.setIndicatorsEnabled(true);
-        build.setLoggingEnabled(true);
-        Picasso.setSingletonInstance(build);
-    }
-
-    @Override
-    protected void attachBaseContext(Context base) {
-        super.attachBaseContext(LocaleManager.setLocale(base));
-    }
-    @Override
-    public void onConfigurationChanged(Configuration newConfig) {
-        super.onConfigurationChanged(newConfig);
-        LocaleManager.setLocale(this);
-    }
-
-}
