--- app/src/main/java/com/shukhrat/wbpvp/ui/feedback_offline/GalleryViewModel.java
@@ -1,19 +1,0 @@
-package com.shukhrat.wbpvp.ui.feedback_offline;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class GalleryViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public GalleryViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is gallery fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
