--- app/src/main/java/com/shukhrat/wbpvp/ui/feedback_online/HomeViewModel.java
@@ -1,19 +1,0 @@
-package com.shukhrat.wbpvp.ui.feedback_online;
-
-import androidx.lifecycle.LiveData;
-import androidx.lifecycle.MutableLiveData;
-import androidx.lifecycle.ViewModel;
-
-public class HomeViewModel extends ViewModel {
-
-    private MutableLiveData<String> mText;
-
-    public HomeViewModel() {
-        mText = new MutableLiveData<>();
-        mText.setValue("This is home fragment");
-    }
-
-    public LiveData<String> getText() {
-        return mText;
-    }
-}
