--- app/src/main/java/com/shukhrat/wbpvp/ui/slideshow/SlideshowFragment.java
@@ -1,55 +1,0 @@
-package com.shukhrat.wbpvp.ui.slideshow;
-
-import android.app.Activity;
-import android.net.ConnectivityManager;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.webkit.WebSettings;
-import android.webkit.WebView;
-import android.webkit.WebViewClient;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import androidx.lifecycle.Observer;
-import androidx.lifecycle.ViewModelProviders;
-
-import com.shukhrat.wbpvp.R;
-
-public class SlideshowFragment extends Fragment {
-
-    private SlideshowViewModel slideshowViewModel;
-    WebView mWebView;
-    ConnectivityManager cm;
-
-    public View onCreateView(@NonNull LayoutInflater inflater,
-                             ViewGroup container, Bundle savedInstanceState) {
-        slideshowViewModel =
-                ViewModelProviders.of(this).get(SlideshowViewModel.class);
-        View root = inflater.inflate(R.layout.fragment_slideshow, container, false);
-
-        mWebView = (WebView) root.findViewById(R.id.webview);
-
-        cm = (ConnectivityManager) getActivity().getSystemService(Activity.CONNECTIVITY_SERVICE);
-        if(cm != null && cm.getActiveNetworkInfo() != null && cm.getActiveNetworkInfo().isConnected()){
-            mWebView.getSettings().setCacheMode(WebSettings.LOAD_DEFAULT);
-        }
-        else{
-            mWebView.getSettings().setCacheMode(WebSettings.LOAD_CACHE_ELSE_NETWORK);
-        }
-
-        mWebView.loadUrl("https://ee.kobotoolbox.org/x/#uCJvSRlS");
-
-        mWebView.setWebViewClient(new WebViewClient());
-
-
-        // Enable Javascript
-        WebSettings webSettings = mWebView.getSettings();
-        webSettings.setJavaScriptEnabled(true);
-
-        return root;
-    }
-}
