+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,56 +7,56 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.shukhrat.wbpvp.ui.feedback_online.FeedbackFragment"
+        android:name="com.piu.ridp.ui.feedback_online.FeedbackFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home">
     </fragment>
 
     <fragment
         android:id="@+id/nav_main"
-        android:name="com.shukhrat.wbpvp.ui.mainpage.MainPageFragment"
+        android:name="com.piu.ridp.ui.mainpage.MainPageFragment"
         tools:layout="@layout/fragment_main_page">
     </fragment>
 
 
     <fragment
         android:id="@+id/nav_feedback"
-        android:name="com.shukhrat.wbpvp.ui.feedback_offline.SavedFeedbackFragment"
+        android:name="com.piu.ridp.ui.feedback_offline.SavedFeedbackFragment"
         tools:layout="@layout/fragment_gallery" />
 
     <fragment
         android:id="@+id/nav_dashboard"
-        android:name="com.shukhrat.wbpvp.ui.slideshow.SlideshowFragment"
+        android:name="com.piu.ridp.ui.slideshow.SlideshowFragment"
         android:label="@string/menu_dashboard"
         tools:layout="@layout/fragment_slideshow" />
     <fragment
         android:id="@+id/nav_news"
-        android:name="com.shukhrat.wbpvp.ui.news.NewsFragment"
+        android:name="com.piu.ridp.ui.news.NewsFragment"
         android:label="@string/news"
         tools:layout="@layout/fragment_news" />
     <fragment
         android:id="@+id/nav_opendata"
-        android:name="com.shukhrat.wbpvp.ui.opendata.OpenDataFragment"
+        android:name="com.piu.ridp.ui.opendata.OpenDataFragment"
         android:label="@string/open_data"
         tools:layout="@layout/fragment_open_data" />
     <fragment
         android:id="@+id/nav_admin"
-        android:name="com.shukhrat.wbpvp.ui.feedback_online.FeedbackFragment"
+        android:name="com.piu.ridp.ui.feedback_online.FeedbackFragment"
         />
 
     <fragment
         android:id="@+id/id_facebook"
-        android:name="com.shukhrat.wbpvp.ui.feedback_online.FeedbackFragment"/>
+        android:name="com.piu.ridp.ui.feedback_online.FeedbackFragment"/>
     <fragment
         android:id="@+id/id_telegram"
-        android:name="com.shukhrat.wbpvp.ui.feedback_online.FeedbackFragment"/>
+        android:name="com.piu.ridp.ui.feedback_online.FeedbackFragment"/>
     <fragment android:id="@+id/id_twitter"
-        android:name="com.shukhrat.wbpvp.ui.feedback_online.FeedbackFragment"/>
+        android:name="com.piu.ridp.ui.feedback_online.FeedbackFragment"/>
     <fragment android:id="@+id/id_share"
-        android:name="com.shukhrat.wbpvp.ui.feedback_online.FeedbackFragment"/>
+        android:name="com.piu.ridp.ui.feedback_online.FeedbackFragment"/>
     <fragment android:id="@+id/logout"
-        android:name="com.shukhrat.wbpvp.ui.feedback_online.FeedbackFragment"/>
+        android:name="com.piu.ridp.ui.feedback_online.FeedbackFragment"/>
     <fragment android:id="@+id/design_bottom_sheet"
-        android:name="com.shukhrat.wbpvp.ui.sliding_panel.BottomSheetFragment"/>
+        android:name="com.piu.ridp.ui.sliding_panel.BottomSheetFragment"/>
 
 </navigation>
